+++ packages/client/src/config/config.ts
@@ -2,11 +2,11 @@ import { Deployment, MediaClientConfig, Tenant } from '../types';
 
 export const mediaClientConfig: Record<Deployment, Record<Tenant, MediaClientConfig>> = {
   local: {
-    davinci: { freeportApiUrl: 'http://localhost:3012' },
-    cerefans: { freeportApiUrl: 'http://localhost:3012' },
+    davinci: { freeportApiUrl: 'http://localhost:3012/' },
+    cerefans: { freeportApiUrl: 'http://localhost:3012/' },
   },
   development: {
-    davinci: { freeportApiUrl: 'https://dev-freeport-api.network-dev.aws.cere.io' },
+    davinci: { freeportApiUrl: 'https://dev-freeport-api.network-dev.aws.cere.io/' },
     cerefans: { freeportApiUrl: 'https://dev-freeport-api-cerefans.network-dev.aws.cere.io/' },
   },
   staging: {