+++ packages/react/src/components/video/VideoPlayer.tsx
@@ -1,29 +1,46 @@
 import './plyr.css';
 import './styles.css';
 
+import clsx from 'clsx';
 import Hls, { HlsConfig } from 'hls.js';
 import Plyr from 'plyr';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { VideoHTMLAttributes, useEffect, useMemo, useRef, useState } from 'react';
 
 interface VideoPlayerProps {
   src: string;
   hlsEnabled?: boolean;
   loader?: HlsConfig['loader'];
+  className?: string;
+  loadingComponent?: React.ReactNode;
+  type?: string;
+  videoOverrides?: VideoHTMLAttributes<HTMLVideoElement>;
 }
 
 const plyrOptions: Plyr.Options = {
   autoplay: true,
 };
 
-export const VideoPlayer = ({ src, hlsEnabled = true, loader }: VideoPlayerProps) => {
+export const VideoPlayer = ({
+  src,
+  hlsEnabled = true,
+  loader = Hls.DefaultConfig.loader,
+  className,
+  loadingComponent,
+  type = 'video/mp4',
+  videoOverrides = { crossOrigin: 'anonymous' },
+}: VideoPlayerProps) => {
   const wrapperRef = useRef<HTMLDivElement>(null);
   const playerRef = useRef<HTMLVideoElement | null>(null);
+  const iosVideoRef = useRef<HTMLVideoElement | null>(null);
   const [isLoading, setIsLoading] = useState<boolean>(true);
 
-  const isSupported = useMemo(() => (hlsEnabled ? Hls.isSupported() : true), [hlsEnabled]);
+  const isVideoSupported = useMemo(() => (hlsEnabled ? Hls.isSupported() : true), [hlsEnabled]);
+  const isIosHlsSupported = useMemo(
+    () => document.createElement('video').canPlayType('application/vnd.apple.mpegurl') !== '',
+    [iosVideoRef],
+  );
 
   useEffect(() => {
-    if (!isSupported) return;
     if (playerRef.current) return;
 
     const videoWrapper = wrapperRef.current;
@@ -38,7 +55,7 @@ export const VideoPlayer = ({ src, hlsEnabled = true, loader }: VideoPlayerProps
 
     const hls = new Hls(hlsOptions);
 
-    if (hlsEnabled) {
+    if (hlsEnabled && isVideoSupported) {
       const updateQuality = (newQuality: number) => {
         if (newQuality === 0) {
           hls.currentLevel = -1;
@@ -61,6 +78,7 @@ export const VideoPlayer = ({ src, hlsEnabled = true, loader }: VideoPlayerProps
         };
         const video = document.createElement('video');
         video.className = 'cere-video';
+        Object.assign(video, videoOverrides);
         playerRef.current = video;
         videoWrapper.appendChild(video);
         hls.attachMedia(video);
@@ -68,17 +86,24 @@ export const VideoPlayer = ({ src, hlsEnabled = true, loader }: VideoPlayerProps
         player.on('canplaythrough', () => setIsLoading(false));
       });
 
-      hls.on(Hls.Events.FRAG_LOADED, () => {
+      hls.on(Hls.Events.FRAG_LOADED, () => setIsLoading(false));
+      hls.loadSource(src);
+    } else if (hlsEnabled && isIosHlsSupported) {
+      const video = iosVideoRef.current;
+      if (!video) return;
+      videoWrapper.appendChild(video);
+      video.src = src;
+      Object.assign(video, videoOverrides);
+      video.addEventListener('canplay', function () {
         setIsLoading(false);
       });
-
-      hls.loadSource(src);
     } else {
       const video = document.createElement('video');
       video.className = 'cere-video';
       playerRef.current = video;
       videoWrapper.appendChild(video);
       video.src = src;
+      Object.assign(video, videoOverrides);
       const player = new Plyr(video, plyrOptions);
       player.on('canplaythrough', () => setIsLoading(false));
       video.addEventListener('error', () => setIsLoading(false));
@@ -93,24 +118,35 @@ export const VideoPlayer = ({ src, hlsEnabled = true, loader }: VideoPlayerProps
       }
       playerRef.current = null;
     };
-  }, [isSupported, loader, src, wrapperRef]);
+  }, [isVideoSupported, loader, src, wrapperRef]);
 
-  if (!isSupported) {
+  if (!isVideoSupported && !isIosHlsSupported) {
     return (
       <div className="cere-video-wrapper flex items-center">
-        <p className="w-full text-center">Your browser does not support HLS.</p>
+        <p className="w-full text-center">Your browser does not support playback of this type.</p>
+      </div>
+    );
+  }
+
+  if (isVideoSupported && !isIosHlsSupported) {
+    return (
+      <div>
+        <div className={clsx('cere-video-wrapper', className)} ref={wrapperRef} />
+
+        {isLoading && (
+          <div className="loading-container">
+            {loadingComponent ? loadingComponent : <div className="loading-container">Loading video...</div>}
+          </div>
+        )}
       </div>
     );
   }
 
   return (
-    <div>
-      <div className="cere-video-wrapper" ref={wrapperRef} />
-      {isLoading && (
-        <div className="loading-container">
-          <>Loading video...</>
-        </div>
-      )}
-    </div>
+    <>
+      <video controls {...videoOverrides}>
+        <source src={src} type={hlsEnabled ? 'application/vnd.apple.mpegurl' : type} />
+      </video>
+    </>
   );
 };